+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - readme: refactor install details & multiple config setup
 - neogit: disable_signs = true otherwise fold signs duplicated
 - community: disable project.nvim in favor of AstroNvim router
+- github: Gists plugin and key maps
 
 ## Changed
 - readme: expand on description and use